[Refactor] OpeningSplashView moved from screensaver to view file
OpeningSplash view seems more appropriate in the view file. imports are fixed after moving the class. Base test should mock the run method of OpeningSplashView as it was mocked before but instead of mocking the whole module only the run method is mocked to properly run the test.
